2024.07.28.03 Commit
작업일 2024.07.28
1. 꼬리치기

+++ Source/HotDogma/LHJ/HD_DragonFSM.h
@@ -136,7 +136,7 @@ public:
 
 	// 타겟과 거리 내적값 추출
 	UFUNCTION()
-	double GetRadianFromCharacter();
+	float GetRadianFromCharacter();
 
 	UFUNCTION()
 	bool ChkCharacterIntoRadian();
